Major Update
I tried to separated out the code into more object oriented.
DelHist is the interface for input files.
DelFill is the interface to read and loop the Delphes tree
DelEvent is the interface to load the event into local variables
->DelTT/DelZJJ/DelWino is the inherited class specified for different process
DelAna is the class to construct the higher level of variables
DelCut is where the cutflow is defined and histogram is filled
HistTool is the tools for histogram manager.

The code is still in beta version, need further validation